# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "DocumentDetail": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "commit_sha": {
            "title": "Commit Sha",
            "type": "string"
          },
          "concept_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "finding_details": {
            "items": {
              "$ref": "#/components/schemas/FindingRead"
            },
            "title": "Finding Details",
            "type": "array"
          },
          "findings": {
            "title": "Findings",
            "type": "integer"
          },
          "frontmatter": {
            "additionalProperties": true,
            "title": "Frontmatter",
            "type": "object"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "synced_at": {
            "format": "date-time",
            "title": "Synced At",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "trust_tier": {
            "title": "Trust Tier",
            "type": "string"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          }
        },
        "required": [
          "path",
          "kind",
          "concept_id",
          "type",
          "title",
          "status",
          "trust_tier",
          "findings",
          "commit_sha",
          "description",
          "frontmatter",
          "body",
          "finding_details",
          "synced_at"
        ],
        "title": "DocumentDetail",
        "type": "object"
      },
      "DocumentSummary": {
        "properties": {
          "commit_sha": {
            "title": "Commit Sha",
            "type": "string"
          },
          "concept_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept Id"
          },
          "findings": {
            "title": "Findings",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "trust_tier": {
            "title": "Trust Tier",
            "type": "string"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          }
        },
        "required": [
          "path",
          "kind",
          "concept_id",
          "type",
          "title",
          "status",
          "trust_tier",
          "findings",
          "commit_sha"
        ],
        "title": "DocumentSummary",
        "type": "object"
      },
      "FindingRead": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "level": {
            "title": "Level",
            "type": "string"
          },
          "line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "level",
          "code",
          "message",
          "line"
        ],
        "title": "FindingRead",
        "type": "object"
      },
...
      "GitHubItemSummary": {
        "properties": {
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "title": "Labels",
            "type": "array"
          },
          "merged_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merged At"
          },
          "number": {
            "title": "Number",
            "type": "integer"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "number",
          "type",
          "title",
          "state",
          "author",
          "labels",
          "url",
          "updated_at",
          "merged_at"
        ],
        "title": "GitHubItemSummary",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KnowledgeReset": {
        "properties": {
          "deleted": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Deleted",
            "type": "object"
          }
        },
        "required": [
          "deleted"
        ],
        "title": "KnowledgeReset",
        "type": "object"
      },
...
      "SyncRunRead": {
        "properties": {
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "requests": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requests"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "stage",
          "error",
          "counts",
          "requests",
          "created_at",
          "finished_at"
        ],
        "title": "SyncRunRead",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkspaceRead": {
        "properties": {
          "branch": {
            "default": "main",
            "maxLength": 255,
            "minLength": 1,
            "title": "Branch",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "embedding_model": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Model"
          },
          "embedding_provider": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Provider"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "index_paths": {
            "items": {
              "type": "string"
            },
            "title": "Index Paths",
            "type": "array"
          },
          "llm_model": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Model"
          },
          "llm_provider": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Provider"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "repository": {
            "examples": [
              "octo-org/demo"
            ],
            "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9-]{0,38})/[A-Za-z0-9._-]{1,100}$",
            "title": "Repository",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "name",
          "repository",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "WorkspaceRead",
        "type": "object"
      },
      "WorkspaceUpdate": {
        "additionalProperties": false,
        "properties": {
          "branch": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch"
          },
          "embedding_model": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Model"
          },
          "embedding_provider": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Provider"
          },
          "index_paths": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Index Paths"
          },
          "llm_model": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Model"
          },
          "llm_provider": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Provider"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "repository": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9-]{0,38})/[A-Za-z0-9._-]{1,100}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repository"
          }
        },
        "title": "WorkspaceUpdate",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Ichnos API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/workspaces/{workspace_id}": {
      "get": {
        "operationId": "getWorkspace",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Workspace not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workspace",
        "tags": [
          "workspaces"
        ]
      },
      "patch": {
        "operationId": "updateWorkspace",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Workspace not found"
          },
          "409": {
            "description": "Name or repository and branch already used"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Workspace",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/workspaces/{workspace_id}/documents": {
      "get": {
        "operationId": "listDocuments",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentSummary"
                  },
                  "title": "Response Listdocuments",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Workspace not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Documents",
        "tags": [
          "sync"
        ]
      }
    },
    "/api/workspaces/{workspace_id}/documents/detail": {
      "get": {
        "description": "One synced document with its frontmatter, body and conformance findings.",
        "operationId": "getDocument",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "maxLength": 500,
              "minLength": 1,
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Workspace or document not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Document",
        "tags": [
          "sync"
        ]
      }
    },
...
    "/api/workspaces/{workspace_id}/github/items": {
      "get": {
        "operationId": "listGitHubItems",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GitHubItemSummary"
                  },
                  "title": "Response Listgithubitems",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Workspace not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Github Items",
        "tags": [
          "sync"
        ]
      }
    },
    "/api/workspaces/{workspace_id}/knowledge": {
      "delete": {
        "description": "Delete all derived knowledge and sync cursors; the next sync rebuilds everything.",
        "operationId": "resetKnowledge",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeReset"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Workspace not found"
          },
          "409": {
            "description": "A sync of this workspace is running"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Workspace Knowledge",
        "tags": [
          "sync"
        ]
      }
    },
...
    "/api/workspaces/{workspace_id}/sync": {
      "post": {
        "operationId": "syncWorkspace",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncRunRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "No GitHub token configured"
          },
          "404": {
            "description": "Workspace not found"
          },
          "409": {
            "description": "A sync of this workspace is already running"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Workspace",
        "tags": [
          "sync"
        ]
      }
    },
    "/api/workspaces/{workspace_id}/sync-runs": {
      "get": {
        "operationId": "listSyncRuns",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SyncRunRead"
                  },
                  "title": "Response Listsyncruns",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Workspace not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Sync Runs",
        "tags": [
          "sync"
        ]
      }
    },
...
  }
}